# daemon/restapi/api_types.go
// Package restapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.1-0.20240325090356-a14414f04fdd DO NOT EDIT.
package restapi

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// Id Primary key identifier
type Id = int64

// IdObject defines model for IdObject.
type IdObject struct {
	// Id Primary key identifier
	Id *Id `json:"id,omitempty"`
}

// IdObjects defines model for IdObjects.
type IdObjects = []IdObject

// Resource defines model for Resource.
type Resource struct {
	Id  *int64  `json:"id,omitempty"`
	Url *string `json:"url,omitempty"`
}

// ResourceWithGroup defines model for ResourceWithGroup.
type ResourceWithGroup struct {
	Group     *string `json:"group,omitempty"`
	GroupId   *int64  `json:"groupId,omitempty"`
	GroupType *string `json:"groupType,omitempty"`
	Id        *int64  `json:"id,omitempty"`
	Url       *string `json:"url,omitempty"`
}

// ResourcesWithGroups defines model for ResourcesWithGroups.
type ResourcesWithGroups = []ResourceWithGroup

// GroupSlug defines model for GroupSlug.
type GroupSlug = string

// GroupType defines model for GroupType.
type GroupType = string

// GroupTypeName defines model for GroupTypeName.
type GroupTypeName = string

// MenuItem defines model for MenuItem.
type MenuItem = string

// UnexpectedError defines model for UnexpectedError.
type UnexpectedError = Error

// PostResourcesWithGroupsJSONRequestBody defines body for PostResourcesWithGroups for application/json ContentType.
type PostResourcesWithGroupsJSONRequestBody = ResourcesWithGroups