# daemon/restapi/openapi.yaml
openapi: 3.0.0
info:
  title: SaveTabs Daemon API
  version: 1.0.0
paths:
  /html/menu:
    get:
      summary: Return the HTML for the Menu
      responses:
        '200':
          $ref: '#/components/responses/HtmlResponse'

  /html/menu/{menuItem}:
    get:
      summary: Return the HTML for the Menu
      parameters:
      - $ref: '#/components/parameters/MenuItem'
      responses:
        '200':
          $ref: '#/components/responses/HtmlResponse'

  /html/browse:
    get:
      summary: Return the HTML for the Browse body, e.g. the list of group types as an HTML list
      responses:
        '200':
          $ref: '#/components/responses/HtmlResponse'

  /html/group-types/{groupTypeName}/groups:
    get:
      summary: Return the list of groups for a group type as HTML
      parameters:
        - $ref: '#/components/parameters/GroupTypeName'
      responses:
        '200':
          $ref: '#/components/responses/HtmlResponse'

  /html/groups/{groupType}/{groupSlug}:
    get:
      summary: Return the list of resources for a group as HTML
      parameters:
        - $ref: '#/components/parameters/GroupType'
        - $ref: '#/components/parameters/GroupSlug'
      responses:
        '200':
          $ref: '#/components/responses/HtmlResponse'

  /resources/with-groups:
    post:
      summary: Adds multiple resources, each with group info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourcesWithGroups'
      responses:
        '201':
          description: The resources were created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IdObjects'
        default:
          $ref: '#/components/responses/UnexpectedError'

  /healthz:
    get:
      summary: Health Check
      description: Check the health status of SaveTabs daemon.
      responses:
        '200':
          description: OK
        '503':
          description: Service Unavailable
  /readyz:
    get:
      summary: Readiness Check
      description: Check if SaveTabs daemon is ready to serve.
      responses:
        '200':
          description: OK
        '503':
          description: Service Unavailable

#  /groups:
#    post:
#      summary: Adds a new group
#      description: >
#        This endpoint when POSTed to adds a new group to the database.
#        The response will simply be the database's `group.id` property.
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              $ref: '#/components/schemas/Group'
#      responses:
#        '201':
#          $ref: '#/components/responses/IdResponse'
#        default:
#          $ref: '#/components/responses/UnexpectedError'
#
#  /groups/{id}:
#    parameters:
#      - $ref: '#/components/parameters/IdParameter'
#
#    get:
#      description: Returns a specific group
#      responses:
#        '200':
#          description: A group object
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/GroupWithId'
#        default:
#          $ref: '#/components/responses/UnexpectedError'
#
#    put:
#      summary: Updates a specific group
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              $ref: '#/components/schemas/GroupWithId'
#      responses:
#        '200':
#          description: The group was updated
#        default:
#          $ref: '#/components/responses/UnexpectedError'
#
#    delete:
#      summary: Deletes a specific group
#      responses:
#        '204':
#          description: The group was deleted
#        default:
#          $ref: '#/components/responses/UnexpectedError'
#
#jsonSchemaDialect: https://spec.openapis.org/oas/3.1/dialect/base
#security:
#  -
#servers:
#  -
#webhooks:
#externalDocs:
#tags:
#  -
components:
  parameters:
    MenuItem:
      name: menuItem
      in: path
      description: Item label for menu option
      required: true
      schema:
        type: string

    GroupTypeName:
      name: groupTypeName
      in: path
      description: Name of group type
      required: true
      schema:
        type: string

    GroupType:
      name: groupType
      in: path
      description: One character group type, e.g. 'G', 'C', 'K', 'T', etc.
      required: true
      schema:
        type: string

    GroupSlug:
      name: groupSlug
      in: path
      description: Slug for the group, i.e. lowercased name with only alphanum and dashes
      required: true
      schema:
        type: string

    IdParameter:
      name: id
      in: path
      description: Primary key of database table
      required: true
      schema:
        type: integer
        minimum: 1
        format: int64

  responses:
    HtmlResponse:
      description: HTML or HTML snippet returned successfully
      content:
        text/html:
          schema:
            type: string
            example: "<div>Hello, HTML!</div>"

    IdResponse:
      description: A generic response containing a primary key Id in JSON format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/IdObject'
    IdsResponse:
      description: A generic response containing multiple primary key Ids in JSON format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/IdObjects'
    UnexpectedError:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Id:
      type: integer
      format: int64
      description: Primary key identifier
    IdObject:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
    IdObjects:
      type: array
      items:
        $ref: '#/components/schemas/IdObject'
    Resource:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: false
        url:
          type: string
          nullable: false
    ResourceWithGroup:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            groupId:
              type: integer
              format: int64
              nullable: false
            group:
              type: string
              nullable: false
            groupType:
              type: string
              nullable: false
    ResourcesWithGroups:
      type: array
      items:
        $ref: '#/components/schemas/ResourceWithGroup'
    Group:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
        type:
          type: string
          enum: [ category, tag, keyword, tab-group ]
    GroupWithId:
      allOf:
        - $ref: '#/components/schemas/Group'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: The unique identifier for the group
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error message