# daemon/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package sqlc

import (
	"database/sql"
)

type Action struct {
	ID   int64          `json:"id"`
	Name sql.NullString `json:"name"`
}

type Group struct {
	ID          int64          `json:"id"`
	Name        string         `json:"name"`
	Type        string         `json:"type"`
	Slug        string         `json:"slug"`
	CreatedTime sql.NullString `json:"created_time"`
	LatestTime  sql.NullString `json:"latest_time"`
	Created     sql.NullInt64  `json:"-"`
	Latest      sql.NullInt64  `json:"latest"`
}

type GroupType struct {
	Type        string         `json:"type"`
	Sort        sql.NullInt64  `json:"sort"`
	Name        sql.NullString `json:"name"`
	Plural      sql.NullString `json:"plural"`
	Description sql.NullString `json:"description"`
}

type GroupsWithCountsView struct {
	ID            int64          `json:"id"`
	ResourceCount int64          `json:"resource_count"`
	Name          string         `json:"name"`
	Type          string         `json:"type"`
	Slug          string         `json:"slug"`
	TypeName      sql.NullString `json:"type_name"`
	TypePlural    sql.NullString `json:"type_plural"`
}

type History struct {
	ID         int64         `json:"id"`
	ForeignID  sql.NullInt64 `json:"foreign_id"`
	DateTime   sql.NullInt64 `json:"date_time"`
	ActionID   sql.NullInt64 `json:"action_id"`
	ActionData []byte        `json:"action_data"`
}

type KeyValue struct {
	ID           int64          `json:"id"`
	ResourceID   sql.NullInt64  `json:"resource_id"`
	Key          sql.NullString `json:"key"`
	Value        sql.NullString `json:"value"`
	CreatedTime  sql.NullString `json:"created_time"`
	ModifiedTime sql.NullString `json:"modified_time"`
	Created      sql.NullInt64  `json:"created"`
	Modified     sql.NullInt64  `json:"modified"`
}

type Resource struct {
	ID          int64          `json:"id"`
	Url         sql.NullString `json:"url"`
	CreatedTime sql.NullString `json:"created_time"`
	VisitedTime sql.NullString `json:"visited_time"`
	Created     sql.NullInt64  `json:"-"`
	Visited     sql.NullInt64  `json:"visited"`
}

type ResourceGroup struct {
	GroupID     sql.NullInt64  `json:"group_id"`
	ResourceID  sql.NullInt64  `json:"resource_id"`
	CreatedTime sql.NullString `json:"created_time"`
	LatestTime  sql.NullString `json:"latest_time"`
	Created     sql.NullInt64  `json:"created"`
	Latest      sql.NullInt64  `json:"latest"`
}

type ResourceGroupIdsView struct {
	ResourceID       sql.NullInt64 `json:"resource_id"`
	GroupIds         string        `json:"group_ids"`
	GroupTypes       string        `json:"group_types"`
	GroupSlugs       string        `json:"group_slugs"`
	GroupNames       string        `json:"group_names"`
	QuotedGroupTypes interface{}   `json:"quoted_group_types"`
	QuotedGroupNames interface{}   `json:"quoted_group_names"`
	QuotedGroupSlugs interface{}   `json:"quoted_group_slugs"`
}

type ResourcesView struct {
	ID               sql.NullInt64  `json:"id"`
	ResourceID       sql.NullInt64  `json:"resource_id"`
	Url              sql.NullString `json:"url"`
	GroupID          int64          `json:"group_id"`
	GroupName        string         `json:"group_name"`
	GroupSlug        string         `json:"group_slug"`
	GroupType        string         `json:"group_type"`
	TypeName         sql.NullString `json:"type_name"`
	Domain           sql.NullString `json:"domain"`
	GroupIds         sql.NullString `json:"group_ids"`
	GroupTypes       sql.NullString `json:"group_types"`
	GroupNames       sql.NullString `json:"group_names"`
	QuotedGroupTypes interface{}    `json:"quoted_group_types"`
	QuotedGroupSlugs interface{}    `json:"quoted_group_slugs"`
	QuotedGroupNames interface{}    `json:"quoted_group_names"`
}

type Var struct {
	ID    int64          `json:"id"`
	Key   string         `json:"key"`
	Value sql.NullString `json:"value"`
}